requirements.txt
Django==1.11.6
django-extensions==1.9.8
django-query-builder==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.3
django-imagekit==4.0.2
